auth_service/requirements.txt
flask
flask-smorest
marshmallow
apispec
python-dotenv
PyMySQL
PyJWT
requests
cryptography
pika
pytest
